fix(runner): keep the carved thread-scope tests inline
`check_no_panics.py`'s `has_cfg_test_module_declaration` only recognises a
FLAT `#[path = "x.rs"]`; a module declared in a non-`mod.rs` file must spell
the directory (`#[path = "loop_driver_host/x.rs"]`), which the regex misses.
The carved file therefore read as production to the delta scan, and its six
fixture `.unwrap()`s failed `Fast deterministic checks`.

Inlining the module is both the fix and the local convention —
`loop_driver_host.rs` already carries four inline `#[cfg(test)]` modules — and
it keeps the three test paths identical (`loop_driver_host::thread_scope_tests::*`).

The scanner gap is pre-existing and latent for the two sibling files declared
the same way (`tests.rs`, `compaction_tests.rs`); neither has ever tripped it
because their panics sit under item-level `#[cfg(test)]` attributes the scanner
does track. Reported rather than fixed here: widening that regex changes a
security-adjacent gate's classification and has baseline implications.

Verified: `--base origin/main --head HEAD`, `--reborn-baseline`, and
`--self-test` all clean; runner test roster unchanged at 259.
